insights.js
This code takes in entries of vape usage sessions, each with a start date, end date, the number of puffs per vape, and the cost per vape. It processes this data to  calculate weekly totals for puffs and spending. It also provides insights on changes in vape usage and spending from week to week, estimating potential yearly savings if spending decreases. The data is then visualized using Chart.jsfor trends.